+++ lib/ts/index.ts
@@ -17,7 +17,7 @@ import AuthHttpRequestFetch from "./fetch";
 import { InputType, RecipeInterface } from "./types";
 import RecipeImplementation from "./recipeImplementation";
 import OverrideableBuilder from "supertokens-js-override";
-import { validateAndNormaliseInputOrThrowError } from "./utils";
+import { getNormalisedUserContext, validateAndNormaliseInputOrThrowError } from "./utils";
 import CookieHandlerReference from "./utils/cookieHandler";
 import WindowHandlerReference from "./utils/windowHandler";
 
@@ -28,7 +28,16 @@ export default class AuthHttpRequest {
         CookieHandlerReference.init(options.cookieHandler);
         WindowHandlerReference.init(options.windowHandler);
         let config = validateAndNormaliseInputOrThrowError(options);
-        const recipeImpl = new OverrideableBuilder(RecipeImplementation()).override(config.override.functions).build();
+        const recipeImpl = new OverrideableBuilder(
+            RecipeImplementation({
+                onHandleEvent: config.onHandleEvent,
+                preAPIHook: config.preAPIHook,
+                postAPIHook: config.postAPIHook,
+                sessionExpiredStatusCode: config.sessionExpiredStatusCode
+            })
+        )
+            .override(config.override.functions)
+            .build();
         AuthHttpRequestFetch.init(config, recipeImpl);
         AuthHttpRequest.axiosInterceptorQueue.forEach(f => {
             f();
@@ -36,37 +45,51 @@ export default class AuthHttpRequest {
         AuthHttpRequest.axiosInterceptorQueue = [];
     }
 
-    static getUserId(): Promise<string> {
-        return AuthHttpRequestFetch.recipeImpl.getUserId(AuthHttpRequestFetch.config);
+    static getUserId(input?: { userContext?: any }): Promise<string> {
+        return AuthHttpRequestFetch.recipeImpl.getUserId({
+            userContext: getNormalisedUserContext(input === undefined ? undefined : input.userContext)
+        });
     }
 
-    static async getAccessTokenPayloadSecurely(): Promise<any> {
-        return AuthHttpRequestFetch.recipeImpl.getAccessTokenPayloadSecurely(AuthHttpRequestFetch.config);
+    static async getAccessTokenPayloadSecurely(input?: { userContext?: any }): Promise<any> {
+        return AuthHttpRequestFetch.recipeImpl.getAccessTokenPayloadSecurely({
+            userContext: getNormalisedUserContext(input === undefined ? undefined : input.userContext)
+        });
     }
 
     static attemptRefreshingSession = async (): Promise<boolean> => {
         return AuthHttpRequestFetch.attemptRefreshingSession();
     };
 
-    static doesSessionExist = () => {
-        return AuthHttpRequestFetch.recipeImpl.doesSessionExist(AuthHttpRequestFetch.config);
+    static doesSessionExist = (input?: { userContext?: any }) => {
+        return AuthHttpRequestFetch.recipeImpl.doesSessionExist({
+            userContext: getNormalisedUserContext(input === undefined ? undefined : input.userContext)
+        });
     };
 
-    static addAxiosInterceptors = (axiosInstance: any) => {
+    static addAxiosInterceptors = (axiosInstance: any, userContext?: any) => {
         if (!AuthHttpRequestFetch.initCalled) {
             // the recipe implementation has not been initialised yet, so add
             // this to the queue and wait for it to be initialised, and then on
             // init call, we add all the interceptors.
             AuthHttpRequest.axiosInterceptorQueue.push(() => {
-                AuthHttpRequestFetch.recipeImpl.addAxiosInterceptors(axiosInstance, AuthHttpRequestFetch.config);
+                AuthHttpRequestFetch.recipeImpl.addAxiosInterceptors({
+                    axiosInstance,
+                    userContext: getNormalisedUserContext(userContext)
+                });
             });
         } else {
-            AuthHttpRequestFetch.recipeImpl.addAxiosInterceptors(axiosInstance, AuthHttpRequestFetch.config);
+            AuthHttpRequestFetch.recipeImpl.addAxiosInterceptors({
+                axiosInstance,
+                userContext: getNormalisedUserContext(userContext)
+            });
         }
     };
 
-    static signOut = () => {
-        return AuthHttpRequestFetch.recipeImpl.signOut(AuthHttpRequestFetch.config);
+    static signOut = (input?: { userContext?: any }) => {
+        return AuthHttpRequestFetch.recipeImpl.signOut({
+            userContext: getNormalisedUserContext(input === undefined ? undefined : input.userContext)
+        });
     };
 }