+++ lib/ts/utils/index.ts
@@ -15,7 +15,14 @@
 
 import NormalisedURLDomain, { isAnIpAddress } from "../normalisedURLDomain";
 import NormalisedURLPath from "../normalisedURLPath";
-import { EventHandler, InputType, NormalisedInputType, RecipeInterface } from "../types";
+import {
+    EventHandler,
+    InputType,
+    NormalisedInputType,
+    RecipeInterface,
+    RecipePostAPIHookFunction,
+    RecipePreAPIHookFunction
+} from "../types";
 import WindowHandlerReference from "../utils/windowHandler";
 import { enableLogging, logDebugMessage } from "../logger";
 
@@ -109,15 +116,18 @@ export function validateAndNormaliseInputOrThrowError(options: InputType): Norma
         cookieDomain = normaliseSessionScopeOrThrowError(options.cookieDomain);
     }
 
-    let preAPIHook = async (context: {
-        action: "SIGN_OUT" | "REFRESH_SESSION";
-        requestInit: RequestInit;
-        url: string;
-    }): Promise<{ url: string; requestInit: RequestInit }> => {
+    let preAPIHook: RecipePreAPIHookFunction = async context => {
         return { url: context.url, requestInit: context.requestInit };
     };
+
     if (options.preAPIHook !== undefined) {
         preAPIHook = options.preAPIHook;
+    }
+
+    let postAPIHook: RecipePostAPIHookFunction = async () => {};
+
+    if (options.postAPIHook !== undefined) {
+        postAPIHook = options.postAPIHook;
     }
 
     let onHandleEvent: EventHandler = () => {};
@@ -146,6 +156,7 @@ export function validateAndNormaliseInputOrThrowError(options: InputType): Norma
         isInIframe,
         cookieDomain,
         preAPIHook,
+        postAPIHook,
         onHandleEvent,
         override
     };
@@ -194,4 +205,12 @@ export function shouldDoInterceptionBasedOnUrl(
             return domain === normalisedCookieDomain;
         }
     }
+}
+
+export function getNormalisedUserContext(userContext?: any): any {
+    if (userContext === undefined) {
+        return {};
+    }
+
+    return userContext;
 }